requirements.txt
requests==2.31.0
openai==1.30.1
diffusers==0.28.2
transformers==4.40.2
torch==2.2.2
torchvision==0.17.2
moviepy==1.0.3
python-dotenv==1.0.1
gtts==2.5.1
syncsdk
requests
ffmpeg-python
numpy==1.23.4
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa==0.9.2 # 
numba
resampy==0.3.1
pydub==0.25.1 
scipy==1.10.1
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml  
joblib==1.1.0
scikit-image==0.19.3
basicsr==1.4.2
facexlib==0.3.0
gradio
gfpgan
av
safetensors
if